requirements_all.txt
absl-py==1.2.0
aiohttp==3.8.1
aiohttp-retry==2.5.2
aiosignal==1.2.0
albumentations==1.2.1
antlr4-python3-runtime==4.9.3
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asttokens==2.0.5
async-timeout==4.0.2
asyncssh==2.8.1
atpublic==3.0.1
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.6.0
bleach==5.0.1
boto3==1.24.34
botocore==1.27.34
bravado==11.0.3
bravado-core==5.17.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
configobj==5.0.6
cryptography==37.0.4
cycler==0.11.0
datasets==2.3.2
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
dill==0.3.5.1
diskcache==5.4.0
distlib==0.3.5
distro==1.7.0
dpath==2.0.6
dulwich==0.20.45
dvc==2.9.5
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
filelock==3.7.1
flake8==4.0.1
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.34.4
fqdn==1.5.1
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grandalf==0.6
grpcio==1.47.0
huggingface-hub==0.8.1
hydra-core==1.2.0
idna==3.3
imageio==2.19.5
importlib-metadata==4.12.0
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isoduration==20.11.0
isort==5.10.1
japanize-matplotlib==1.1.3
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.8
jsonpointer==2.3
jsonref==0.2
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kaggle==1.5.12
kiwisolver==1.4.4
mailchecker==4.1.18
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
-e git+https://github.com/k-washi/azure-exp.git@4ff9d5525e9f82cc62e7e81d08d59c88dcf1a4c5#egg=mlexpenv
monotonic==1.6
msgpack==1.0.4
multidict==6.0.2
multiprocess==0.70.13
mypy==0.971
mypy-extensions==0.4.3
nanotime==0.5.2
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
neptune-client==0.14.3
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.1
oauthlib==3.2.0
omegaconf==2.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.52
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyDeprecate==0.3.2
pydot==1.4.2
pyflakes==2.4.0
pygit2==1.9.2
Pygments==2.12.0
pygtrie==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-benedict==0.25.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-fsutil==0.6.1
python-slugify==6.1.2
pytorch-lightning==1.6.5
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
qudida==0.0.4
regex==2022.7.9
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==12.5.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.1
scikit-plot==0.3.7
scipy==1.8.1
scmrepo==0.0.7
seaborn==0.11.2
Send2Trash==1.8.0
shortuuid==1.0.9
shtab==1.5.5
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
swagger-spec-validator==2.7.4
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchmetrics==0.9.2
torchsummary==1.5.1
torchvision==0.11.3+cu113
tornado==6.2
tox==3.25.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
typing_extensions==4.3.0
uri-template==1.2.0
urllib3==1.26.10
virtualenv==20.15.1
voluptuous==0.13.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.1
xmltodict==0.13.0
xxhash==3.0.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.8.1
